+++ src/kernel/mod.rs
@@ -1,4 +1,5 @@
 pub mod clib;
 pub mod processmgr;
+pub mod rng;
 pub mod sysmem;
 pub mod threadmgr;