+++ src/kernel/rng.rs
@@ -1,0 +1,4 @@
+#[link(kind = "static", name = "SceLibKernel_stub")]
+extern "C" {
+    pub fn sceKernelGetRandomNumber(output: *mut ::void, size: u32);
+}